more data seeded

+++ fa22team31finalproject/Models/StockPortfolio.cs
@@ -7,11 +7,11 @@ namespace fa22team31finalproject.Models
 {
     public class StockPortfolio
     {
-        public Int32 StockPortfolioID { get; set;
-        }
+        public Int32 StockPortfolioID { get; set; }
+
         [Display(Name = "Account Number:")]
         [Required(ErrorMessage = "Account Number is required.")]
-        public Int32 AccountNumber { get; set; }
+        public Int64 AccountNumber { get; set; }
 
         [Display(Name = "Account Name:")]
         [Required(ErrorMessage = "Account Name is required.")]
